Move traits to their own modules
